+++ packages/remme-atomic-swap/src/models/SwapInfoDto.ts
@@ -1,5 +1,7 @@
+import { SwapState } from "./SwapState";
+
 export class SwapInfoDto {
-    public state: State;
+    public state: SwapState;
     public receiver_address: string;
     public amount: number;
     public email_address_encrypted_optional: string;
@@ -10,13 +12,4 @@ export class SwapInfoDto {
     public sender_address: string;
     public sender_address_non_local: string;
     public swap_id: string;
-}
-
-export enum State {
-    EMPTY = "EMPTY",
-    OPENED = "OPENED",
-    SECRET_LOCK_PROVIDED = "SECRET_LOCK_PROVIDED",
-    APPROVED = "APPROVED",
-    CLOSED = "CLOSED",
-    EXPIRED = "EXPIRED",
 }