+++ packages/remme-blockchain-info/src/models/Receipts.ts
@@ -1,4 +1,3 @@
 export interface ReceiptList {
     data: any;
-    link: string;
 }