+++ packages/remme-web-socket/src/models/index.ts
@@ -1,3 +1,3 @@
-export { BatchStateUpdateDto, BatchStatusesDto, Statuses, BatchStatuses } from "./BatchStateUpdateDto";
+export { BatchStateUpdateDto, BatchInfo, BatchInfoDto, BatchStatus } from "./BatchStateUpdateDto";
 export { ErrorMessage } from "./ErrorMessage";
 export { ErrorFromEvent } from "./ErrorFromEvent";